Feat(#3): Calendar Entity 작성
