--- domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/calendar/Calendar.java
@@ -1,17 +1,0 @@
-package com.wypl.jpacalendardomain.calendar;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class Calendar {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private String description;
-    private Long ownerId;
-    private boolean isShared;
-}
